refactor(desktop): move to own module
